broken: start event emitter layout
