+++ src/main/emitter.js
@@ -1,0 +1,5 @@
+import EventEmitter from 'events'
+
+const emitter = new EventEmitter()
+
+export default emitter
