+++ src/main/scanForApps.js
